refactor: Commands to their own pkgs
