+++ cmd/meroxa/root/auth/login.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package root
+package auth
 
 import (
 	"context"
