+++ cmd/meroxa/root/completion.go
@@ -1,4 +1,4 @@
-package deprecated
+package root
 
 import (
 	"fmt"
