+++ cmd/meroxa/root/open/open_test.go
@@ -1,4 +1,4 @@
-package deprecated_test
+package open_test
 
 import (
 	"bytes"
